requirements.txt
ddgs
google-genai
python-dotenv
rich
